+++ CHANGELOG.md
@@ -1,3 +1,10 @@
 ## 0.0.1
 
-* TODO: Describe initial release.
+* Initial release.
+
+## 0.0.2
+
+* State logic separated from UI logic using flutter_hooks.
+* Added support for row height and column width resizing, 
+* Added support for row and column selection
+* Code cleanup